Move base tool data infrastructure into tool_util package.
Keep the refgenie stuff in Galaxy's app package because it brings in additional dependencies.

This is a small part of a larger ongoing project to cleanup tool shed dependencies and allow tool shed repository installations with minimal dependencies.
